pyproject.toml
[tool.poetry]
name = "ASRD"
version = "0.1.0"
description = ""
authors = ["Mikhail Borisov <travka_mx@mail.ru>"]
license = "Unlicensed"

[tool.poetry.dependencies]
python = "^3.9"
fsspec = "^0.8.7"
Flask = "^1.1.2"
python-dotenv = "^0.15.0"
Flask-BasicAuth = "^0.2.0"
gunicorn = "^20.1.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^2.7.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
